remote snapshot sense volume status
1. add volume attr class
2. snapshot and replay do according to volume attr
3. remote nouse function declaration in journal replayer
4. fixbug in volume add volume status
5. rewrite replica replay use in memory consumer mark

+++ src/sg_client/control/control_volume.h
@@ -47,7 +47,7 @@ public:
 private:
     bool execute_cmd(const std::string& command, std::string& result);
 
-    bool create_volume(const std::string& volume_id, int size,
+    bool create_volume(const std::string& volume_id, size_t size,
             const std::string& device);
     std::string get_target_iqn(const std::string& volume_id);
     bool generate_config(const std::string& volume_id,
